// vega-ms-web/src/pages/department/reallocation/list/data.d.ts
export interface TableListItem {
	key: number;
	disabled?: boolean;
	href: string;
	avatar: string;
	name: string;
	owner: string;
	desc: string;
	callNo: number;
	status: number;
	updatedAt: Date;
	createdAt: Date;
	progress: number;
}

export interface TableListPagination {
	total: number;
	pageSize: number;
	current: number;
}

export interface TableListData {
	list: TableListItem[];
	pagination: Partial<TableListPagination>;
}

export interface TableListParams {
	sorter?: string;
	status?: string;
	name?: string;
	desc?: string;
	key?: number;
	pageSize?: number;
	currentPage?: number;
}